Refactor TrainingJob
