+++ application/backend/app/core/jobs/models/training_job.py
@@ -12,7 +12,7 @@ from app.core.jobs import Job, JobParams, JobType
 from app.models import Task
 
 
-class TrainingParams(JobParams):
+class TrainingJobParams(JobParams):
     job_id: UUID | None = None
     project_id: UUID | None = None
     model_architecture_id: str
@@ -29,7 +29,7 @@ class TrainingJob(ProjectJob):
     job_type: Literal[JobType.TRAIN] = JobType.TRAIN
     log_dir: Path
     data_dir: Path
-    params: TrainingParams
+    params: TrainingJobParams
 
     def model_post_init(self, _: Any) -> None:
         self.params.job_id = self.id
