+++ application/backend/tests/unit/services/training/test_otx_trainer.py
@@ -11,6 +11,7 @@ from otx.tools.converter import GetiConfigConverter
 from otx.types.export import OTXExportFormatType
 from otx.types.precision import OTXPrecisionType
 
+from app.core.jobs.models import TrainingJobParams
 from app.core.run import ExecutionContext
 from app.models import DatasetItemAnnotationStatus, DatasetItemSubset, Task, TaskType, TrainingStatus
 from app.models.training_configuration.configuration import (
@@ -28,7 +29,6 @@ from app.services import (
     TrainingConfigurationService,
 )
 from app.services.base_weights_service import BaseWeightsService
-from app.services.training.models import TrainingParams
 from app.services.training.otx_trainer import OTXTrainer, TrainingDependencies
 from app.services.training.subset_assignment import (
     DatasetItemWithLabels,
@@ -128,7 +128,7 @@ class TestOTXTrainerPrepareWeights:
     ):
         """Test preparing weights when no parent model revision ID is provided."""
         # Arrange
-        training_params = TrainingParams(
+        training_params = TrainingJobParams(
             model_architecture_id="Object_Detection_YOLOX_S",
             task=Task(task_type=TaskType.DETECTION),
             parent_model_revision_id=None,
@@ -156,7 +156,7 @@ class TestOTXTrainerPrepareWeights:
         # Arrange
         project_id = uuid4()
         parent_model_revision_id = uuid4()
-        training_params = TrainingParams(
+        training_params = TrainingJobParams(
             project_id=project_id,
             model_architecture_id="Object_Detection_YOLOX_S",
             task=Task(task_type=TaskType.DETECTION),
@@ -184,7 +184,7 @@ class TestOTXTrainerPrepareWeights:
         # Arrange
         project_id = uuid4()
         parent_model_revision_id = uuid4()
-        training_params = TrainingParams(
+        training_params = TrainingJobParams(
             project_id=project_id,
             model_architecture_id="Object_Detection_YOLOX_S",
             task=Task(task_type=TaskType.DETECTION),
@@ -205,7 +205,7 @@ class TestOTXTrainerPrepareWeights:
     ):
         """Test that ValueError is raised when parent model revision ID is provided without project ID."""
         # Arrange
-        training_params = TrainingParams(
+        training_params = TrainingJobParams(
             model_architecture_id="Object_Detection_YOLOX_S",
             task=Task(task_type=TaskType.DETECTION),
             parent_model_revision_id=uuid4(),
@@ -225,7 +225,7 @@ class TestOTXTrainerPrepareTrainingConfiguration:
         # Arrange
         project_id = uuid4()
         parent_model_revision_id = uuid4()
-        training_params = TrainingParams(
+        training_params = TrainingJobParams(
             project_id=project_id,
             model_architecture_id="Object_Detection_YOLOX_S",
             task=Task(task_type=TaskType.DETECTION),
@@ -499,7 +499,7 @@ class TestOTXTrainerPrepareModel:
         project_id = uuid4()
         model_id = uuid4()
         model_architecture_id = "Custom_Image_Classification_EfficientNet-B0"
-        training_params = TrainingParams(
+        training_params = TrainingJobParams(
             model_id=model_id,
             project_id=project_id,
             model_architecture_id=model_architecture_id,
@@ -724,7 +724,7 @@ class TestOTXTrainerStoreModelArtifacts:
         project_id = uuid4()
         model_id = uuid4()
 
-        training_params = TrainingParams(
+        training_params = TrainingJobParams(
             model_id=model_id,
             project_id=project_id,
             model_architecture_id="Object_Detection_YOLOX_S",
